refactor: improve build_url
Improved typings of the BuildUrlFunction and
move the Url type to the build_url file
